usm: Unify SSL implementations
